Fix warning
Updating votes used to be in the render function which caused a warning
because the rating would update every time the component rendered.  This
uneccessarily flooded redux with several state changes.  Move logic to
component did mount so rating is tallied after votes are fetched rather
than tallying every time the component is rendered.